manque strlcat, itoa, strsplit

+++ ft_strnstr.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strnstr.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pthan <marvin@42.fr>                       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/20 15:30:15 by pthan             #+#    #+#             */
+/*   Updated: 2018/11/20 19:36:08 by pthan            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (needle == 0)
@@ -13,7 +25,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 		j = 0;
 		if (haystack[i] == needle[j])
 		{
-			while (haystack[i + j ] == needle[j] && needle[j] && i + j < len)
+			while (haystack[i + j] == needle[j] && needle[j] && i + j < len)
 				j++;
 			if (needle[j] == '\0')
 				return ((char *)(haystack + i));
