+++ ft_strrchr.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strrchr.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pthan <marvin@42.fr>                       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/20 09:52:07 by pthan             #+#    #+#             */
+/*   Updated: 2018/11/20 09:54:03 by pthan            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 char	*ft_strrchr(const char *s, int c)
 {
 	char	*tmp;
-	int	n;
+	size_t	n;
 
 	tmp = (char*)s;
 	n = ft_strlen(tmp);
@@ -14,7 +26,7 @@ char	*ft_strrchr(const char *s, int c)
 	while (n > 0)
 	{
 		n--;
-		if (*(tmp + n) == (char) (c))
+		if (*(tmp + n) == (char)(c))
 			return (tmp + n);
 	}
 	return (NULL);
